Fix STM32 gitignore and clean tracked files

+++ STM32N6-GettingStarted-ObjectDetection/Application/STM32N6570-DK/Inc/zip_postprocess.h
@@ -18,14 +18,14 @@
 /* Dimensões do density map de saída do modelo ZIP */
 #define ZIP_OUT_H  16
 #define ZIP_OUT_W  16
-#define ZIP_OUT_SIZE (ZIP_OUT_H * ZIP_OUT_W)  /* 256 floats */
+#define ZIP_OUT_SIZE (ZIP_OUT_H * ZIP_OUT_W)  /* 256 int8 values */
 
 /**
  * Conta pessoas a partir do density map de saída do modelo ZIP.
  *
- * @param density_map  Ponteiro para o buffer de saída da rede (256 floats)
+ * @param density_map  Ponteiro para o buffer de saída da rede (256 int8)
  * @return             Número estimado de pessoas (inteiro)
  */
-uint32_t ZIP_Postprocess_GetCount(const float *density_map);
+uint32_t ZIP_Postprocess_GetCount(const int8_t *density_map);
 
 #endif /* ZIP_POSTPROCESS_H */
